Move rectify code into concern
